remove redundant abstraction
