+++ ethabi/src/event_param.rs
@@ -1,6 +1,6 @@
 //! Event param specification.
 
-use super::ParamType;
+use {ParamType};
 
 /// Event param specification.
 #[derive(Debug, Clone, PartialEq, Deserialize)]
@@ -17,8 +17,7 @@ pub struct EventParam {
 #[cfg(test)]
 mod tests {
 	use serde_json;
-	use super::EventParam;
-	use spec::ParamType;
+	use {EventParam, ParamType};
 
 	#[test]
 	fn event_param_deserialization() {
@@ -27,7 +26,7 @@ mod tests {
 			"type": "address",
 			"indexed": true
 		}"#;
-		
+
 		let deserialized: EventParam = serde_json::from_str(s).unwrap();
 
 		assert_eq!(deserialized, EventParam {
