+++ ethabi/src/param.rs
@@ -1,20 +1,34 @@
-use spec::{self, ParamType};
+//! Function param.
+use ParamType;
 
 /// Function param.
-#[derive(Clone, Debug, PartialEq)]
+#[derive(Debug, Clone, PartialEq, Deserialize)]
 pub struct Param {
 	/// Param name.
 	pub name: String,
-	/// Param kind.
+	/// Param type.
+	#[serde(rename="type")]
 	pub kind: ParamType,
 }
 
-impl From<spec::Param> for Param {
-	fn from(param: spec::Param) -> Self {
-		Param {
-			name: param.name,
-			kind: param.kind,
-		}
+#[cfg(test)]
+mod tests {
+	use serde_json;
+	use super::Param;
+	use spec::ParamType;
+
+	#[test]
+	fn param_deserialization() {
+		let s = r#"{
+			"name": "foo",
+			"type": "address"
+		}"#;
+
+		let deserialized: Param = serde_json::from_str(s).unwrap();
+
+		assert_eq!(deserialized, Param {
+			name: "foo".to_owned(),
+			kind: ParamType::Address,
+		});
 	}
 }
-
