--- ethabi/src/spec/mod.rs
@@ -1,18 +1,0 @@
-//! Contract interface specification.
-
-mod constructor;
-mod event;
-mod event_param;
-mod interface;
-mod operation;
-mod param;
-pub mod param_type;
-
-pub use self::constructor::Constructor;
-pub use self::event::Event;
-pub use self::event_param::EventParam;
-pub use self::interface::Interface;
-pub use self::operation::Operation;
-pub use self::param::Param;
-pub use self::param_type::ParamType;
-
